now have basic impl of the main tasks in ingest.py
With a bit of hard coding here and there we now have the following tasks in place: find the orbital period/pe & phase fold the light-curve, interpolate the folded light-curve over 1024 data points, get the ML model to make predictions for JKTEBOP params and finally generate JKTEBOP in & dat files for processing.

It all largely works. There's some detail work to do but the basic functionality is largely in place.