now sets primary_epoch from most prominent eclipse
Implemented an algorithm for finding the primary_epoch.  It uses scipy.find_peaks() to find significant peaks (>1*sigma).  The primary_epoch is given the time of the most prominent of these peaks.

Also moved the plotting of the light-curve to happen after finding the primary_epoch so that it can be marked on the plot.