improved command line arguments & MAST interaction
Changes -s/--system argument to -t/--target
Added -s/--sectors optional argument which takes an integer list of sectors
Added hidden arguments for -m/--mission and -a/--author which are currently fixed at TESS & SPOC
Reordered arguments and updated readme

Updated the MAST interaction to include searching on sectors and to read the arguments for the mission and author. Also now downloads directly into the iterated lcs (so it reflects any specified sectors). Commented out the block that loads the lcs from files in the cache but left in place as it's useful.