added support for ingest configured by json file
ingest.py now supports the -f/--file command line argument which instructs it to pick up its config from the specified json file.  Either the -f or the -t argument *must* be specified (not both). The previous -f for flux argument has been renamed -fl.

Overriding of file based config is supported.  While working from a file, any command line argument are still picked up and will be used to override the equivalent values in the file (if they're not the default).

The use of the json file will allow more complex config to be written and opens up the possibility of configs that would be very difficult/fiddly to express on the command line. It's also persisted so the pipeline config can be retained as a project asset.