fix for #4 - added an --exptime filter argument
Now allows the exptime paramter of lk.search_lightcurve() to be set via the ingest.py arguments or json file.  Enables a search to filter on assets matching long (10 - 30 min), short (1 or 2 min) or fast (20 sec) cadence.

Is a fix for #4 where without this filter I was retrieving a fast cadence asset for V455 Aur sector 60 which was breaking the pipeline (its meta dict had its cadence at 120 s).