added support for specifying sys_name in json
Previously sys_name was a synonym for target.  Now it can be specified directly (defaults to target if not specified).  This allows user to specify target with an object ID and the sys_name with a readable name for comments and titles.

The file/folder names are now more clearly derived from sys_name.

Included the object ID in the narrative that preceeds the processing of each sector.