resolved #2 - interpolation failure KX Cnc
KX Cnc highlighted an issue with the interpolation of the folded light-curve.  If the light-curve is sparse/has incomplete coverage there may be significant gaps in the folded light-curve at the start and/or end.  If this is the case the interpolation will fail in the -0.25 or +0.75 region as it needs adjascent values on both sides to interpolate from.

The solutions is to move the interpolation code into a new function get_reduced_folded_lc().  Within this I've implemented some logic which extends both the top and bottom of the phase space by copying rows from the opposite end.  The gaps are still there but there will be some adjascent data for the interpolation.

Also adjusted the folded plot markers so that it's easier to see gaps in the underlying phase space.