added auto-poly capability
In addtion to the existing "date_range", poly configs now support an alternative "gap_threshold" config option. The options are exclusive, if a "date_range" item is present it takes priority. When the "gap_threshold" option is used,  one or more date_ranges are calculated from the light-curve data split on gaps in the data greater than the threshold (in days). Each date range is used to build a poly instruction, so a single "gap_threshold" config may generate 1 or more poly instructions (if no gaps, there will be one poly instruction for the whole light-curve).

You can't apply both types of poly to the same light-curve.  Once a poly of one type has been triggered for a light-curve all polies of the other type are ignored.

Polies are processed in order, so the expected set up is any "date_range" polies will be listed before a final, optional "gap_threshold" poly. In this setup, for each light-curve, the "date_range" polies will be tested and applied where they overlap on the time axis. If none are applied the "gap_threshold" poly will be triggered and applied.