added fitting_params overrides to json
Now allows for jktebop fitting parameters to be set to specific values using the "fitting_params" item (dict) of the target json file.  Add a key matching any of the tokens in the template and the accompanying value will be used to override any default, estimated or calculated value.

Particularly useful for Limb Darkening params which are hard coded otherwise.  Added tokens to control whether the LD coefs are fitted. Changed the default values to some a simple set of quad params.

Updated the cw_eri.json with LD, L3 and qphot fitting_params to fully demonstrate their use.