added Estimator class wrapper for ML model
ingest now interacts with the ML model via the Estimator class.  This abstracts the nuts and bolts of the model, especially the ordering and scaling of the features.  The predict function returns a pandas DataFrame with a column for each named feature.

It's expected that an Estimator class will be provided/updated alongside any model so that the ingest class need not be aware of any structural changes.