resolves #23 - negative L3 breaking JKTEBOP
We now catch negative L3 values and update to zero prior to writing out the JKTEBOP task3 in file.  Below is an example of the messaging output for this.

Writing JKTEBOP task3 in file to 'cw_eri_s0004.in'
Negative L3 values are not supported for JKTEBOP in files. Updating L3 from -0.08 to zero.
Writing 13841 rows(s) for columns ['time', 'delta_mag', 'delta_mag_err'] to 'cw_eri_s0004.dat'