resolves #24 - no longer use the -f switch
The args.file argument is now positional and should not be preceeded with a -f or --file.  It's compulsory, and directly indicates the target json file to load, parse and process unless the -n arg is present in which case the file is (over)written with a new, default target config.