updated model based on constrained tset
This model trained on training set with restricted mass range (0.8 to 2.6 M_sun) and period (2.0 to 1.5 d) to allow finer increments in the initial masses, inc and omega parameters.  The model now getting good fit on CW Eri rA+rB, k, J, inc params. The ecosw & esinw not so good but the overall set of estimates is good enough to pass straight through to JKTEBOP to get a good fit.